Fixes a bug that prevented processing of files > 4 GB. Thanks to Stephan X. Esterhuizen for sending the patch.
git-svn-id: https://svn.code.sf.net/p/gnss-sdr/code/trunk@415 64b25241-fba3-4117-9849-534c7e92360d

+++ src/algorithms/libs/gnss_sdr_valve.cc
@@ -39,7 +39,7 @@
 using google::LogMessage;
 
 gnss_sdr_valve::gnss_sdr_valve (size_t sizeof_stream_item,
-        int nitems,
+        unsigned long long nitems,
         gr::msg_queue::sptr queue) : gr::sync_block("valve",
                 gr::io_signature::make(1, 1, sizeof_stream_item),
                 gr::io_signature::make(1, 1, sizeof_stream_item) ),
@@ -49,7 +49,7 @@ gnss_sdr_valve::gnss_sdr_valve (size_t sizeof_stream_item,
 
 
 boost::shared_ptr<gr::block> gnss_sdr_make_valve (size_t sizeof_stream_item,
-        int nitems,
+        unsigned long long nitems,
         gr::msg_queue::sptr queue)
 {
     return boost::shared_ptr<gnss_sdr_valve> (new gnss_sdr_valve (sizeof_stream_item, nitems, queue));
@@ -69,9 +69,8 @@ int gnss_sdr_valve::work (int noutput_items,
             delete cmf;
             return -1;	// Done!
         }
-    unsigned n = std::min(d_nitems - d_ncopied_items, noutput_items);
-    if (n == 0)
-        return 0;
+    unsigned long long n = std::min(d_nitems - d_ncopied_items, noutput_items);
+    if (n == 0) return 0;
     memcpy (output_items[0], input_items[0], n * input_signature()->sizeof_stream_item(0));
     d_ncopied_items += n;
     return n;
