+++ src/algorithms/libs/gnss_sdr_valve.h
@@ -40,7 +40,7 @@
 
 
 boost::shared_ptr<gr::block> gnss_sdr_make_valve (size_t sizeof_stream_item,
-        int nitems,
+        unsigned long long nitems,
         gr::msg_queue::sptr queue);
 /*!
  * \brief Implementation of a GNU Radio block that sends a STOP message to the
@@ -49,13 +49,13 @@ boost::shared_ptr<gr::block> gnss_sdr_make_valve (size_t sizeof_stream_item,
 class gnss_sdr_valve : public gr::sync_block
 {
     friend boost::shared_ptr<gr::block> gnss_sdr_make_valve(size_t sizeof_stream_item,
-                                                            int nitems,
-                                                            gr::msg_queue::sptr queue);
+            unsigned long long nitems,
+            gr::msg_queue::sptr queue);
     gnss_sdr_valve (size_t sizeof_stream_item,
-                    int nitems,
-                    gr::msg_queue::sptr queue);
-    int	d_nitems;
-    int	d_ncopied_items;
+            unsigned long long nitems,
+            gr::msg_queue::sptr queue);
+    unsigned long long	d_nitems;
+    unsigned long long	d_ncopied_items;
     gr::msg_queue::sptr d_queue;
 
 public:
